+++ packages/@webex/plugin-meetings/src/meeting/util.ts
@@ -203,7 +203,11 @@ const MeetingUtil = {
         meeting.reconnectionManager.cleanUp();
       })
       .then(() => meeting.stopKeepAlive())
-      .then(() => meeting.updateLLMConnection());
+      .then(() => {
+        if (meeting.config?.enableAutomaticLLM) {
+          meeting.updateLLMConnection();
+        }
+      });
   },
 
   disconnectPhoneAudio: (meeting, phoneUrl) => {