+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/utils.js
@@ -57,7 +57,8 @@ describe('plugin-meetings', () => {
     });
 
     describe('#cleanup', () => {
-      it('do clean up on meeting object', async () => {
+      it('do clean up on meeting object with LLM enabled', async () => {
+        meeting.config = {enableAutomaticLLM : true};
         await MeetingUtil.cleanUp(meeting);
         assert.calledOnce(meeting.cleanupLocalStreams);
         assert.calledOnce(meeting.closeRemoteStreams);
@@ -68,6 +69,37 @@ describe('plugin-meetings', () => {
         assert.calledOnce(meeting.reconnectionManager.cleanUp);
         assert.calledOnce(meeting.stopKeepAlive);
         assert.calledOnce(meeting.updateLLMConnection);
+        assert.calledOnce(meeting.breakouts.cleanUp);
+        assert.calledOnce(meeting.simultaneousInterpretation.cleanUp);
+      });
+
+      it('do clean up on meeting object with LLM disabled', async () => {
+        meeting.config = {enableAutomaticLLM : false};
+        await MeetingUtil.cleanUp(meeting);
+        assert.calledOnce(meeting.cleanupLocalStreams);
+        assert.calledOnce(meeting.closeRemoteStreams);
+        assert.calledOnce(meeting.closePeerConnections);
+
+        assert.calledOnce(meeting.unsetRemoteStreams);
+        assert.calledOnce(meeting.unsetPeerConnections);
+        assert.calledOnce(meeting.reconnectionManager.cleanUp);
+        assert.calledOnce(meeting.stopKeepAlive);
+        assert.notCalled(meeting.updateLLMConnection);
+        assert.calledOnce(meeting.breakouts.cleanUp);
+        assert.calledOnce(meeting.simultaneousInterpretation.cleanUp);
+      });
+
+      it('do clean up on meeting object with no config', async () => {
+        await MeetingUtil.cleanUp(meeting);
+        assert.calledOnce(meeting.cleanupLocalStreams);
+        assert.calledOnce(meeting.closeRemoteStreams);
+        assert.calledOnce(meeting.closePeerConnections);
+
+        assert.calledOnce(meeting.unsetRemoteStreams);
+        assert.calledOnce(meeting.unsetPeerConnections);
+        assert.calledOnce(meeting.reconnectionManager.cleanUp);
+        assert.calledOnce(meeting.stopKeepAlive);
+        assert.notCalled(meeting.updateLLMConnection);
         assert.calledOnce(meeting.breakouts.cleanUp);
         assert.calledOnce(meeting.simultaneousInterpretation.cleanUp);
       });
@@ -369,7 +401,7 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {
           reachability: 'reachability',
           roapMessage: 'roapMessage',
@@ -401,6 +433,7 @@ describe('plugin-meetings', () => {
             mediaConnections: 'mediaConnections',
           },
         });
+        parseLocusJoinSpy.restore();
       });
 
       it('#Should call meetingRequest.joinMeeting with breakoutsSupported=true when passed in as true', async () => {
@@ -411,7 +444,7 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {
           breakoutsSupported: true,
         });
@@ -420,6 +453,7 @@ describe('plugin-meetings', () => {
         const parameter = meeting.meetingRequest.joinMeeting.getCall(0).args[0];
 
         assert.equal(parameter.breakoutsSupported, true);
+        parseLocusJoinSpy.restore();
       });
 
       it('#Should call meetingRequest.joinMeeting with liveAnnotationSupported=true when passed in as true', async () => {
@@ -430,7 +464,7 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {
           liveAnnotationSupported: true,
         });
@@ -439,6 +473,7 @@ describe('plugin-meetings', () => {
         const parameter = meeting.meetingRequest.joinMeeting.getCall(0).args[0];
 
         assert.equal(parameter.liveAnnotationSupported, true);
+        parseLocusJoinSpy.restore();
       });
 
       it('#Should call meetingRequest.joinMeeting with locale=en_UK, deviceCapabilities=["TEST"] when they are passed in as those values', async () => {
@@ -449,7 +484,7 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {
           locale: 'en_UK',
           deviceCapabilities: ['TEST'],
@@ -460,6 +495,7 @@ describe('plugin-meetings', () => {
 
         assert.equal(parameter.locale, 'en_UK');
         assert.deepEqual(parameter.deviceCapabilities, ['TEST']);
+        parseLocusJoinSpy.restore();
       });
 
       it('#Should call meetingRequest.joinMeeting with preferTranscoding=false when multistream is enabled', async () => {
@@ -473,7 +509,7 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {});
 
         assert.calledOnce(meeting.meetingRequest.joinMeeting);
@@ -481,6 +517,7 @@ describe('plugin-meetings', () => {
 
         assert.equal(parameter.inviteeAddress, 'meetingJoinUrl');
         assert.equal(parameter.preferTranscoding, false);
+        parseLocusJoinSpy.restore();
       });
 
       it('#Should fallback sipUrl if meetingJoinUrl does not exists', async () => {
@@ -493,13 +530,14 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {});
 
         assert.calledOnce(meeting.meetingRequest.joinMeeting);
         const parameter = meeting.meetingRequest.joinMeeting.getCall(0).args[0];
 
         assert.equal(parameter.inviteeAddress, 'sipUri');
+        parseLocusJoinSpy.restore();
       });
 
       it('#Should fallback to meetingNumber if meetingJoinUrl/sipUrl  does not exists', async () => {
@@ -512,7 +550,7 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {});
 
         assert.calledOnce(meeting.meetingRequest.joinMeeting);
@@ -520,6 +558,7 @@ describe('plugin-meetings', () => {
 
         assert.isUndefined(parameter.inviteeAddress);
         assert.equal(parameter.meetingNumber, 'meetingNumber');
+        parseLocusJoinSpy.restore();
       });
 
       it('should pass in the locusClusterUrl from meetingInfo', async () => {
@@ -533,19 +572,20 @@ describe('plugin-meetings', () => {
           getWebexObject: sinon.stub().returns(webex),
         };
 
-        MeetingUtil.parseLocusJoin = sinon.stub();
+        const parseLocusJoinSpy = sinon.stub(MeetingUtil, 'parseLocusJoin');
         await MeetingUtil.joinMeeting(meeting, {});
 
         assert.calledOnce(meeting.meetingRequest.joinMeeting);
         const parameter = meeting.meetingRequest.joinMeeting.getCall(0).args[0];
 
         assert.equal(parameter.locusClusterUrl, 'locusClusterUrl');
+        parseLocusJoinSpy.restore();
       });
     });
 
     describe('joinMeetingOptions', () => {
       it('sends client events correctly', async () => {
-        MeetingUtil.joinMeeting = sinon.stub().rejects({});
+        const joinMeetingSpy = sinon.stub(MeetingUtil, 'joinMeeting').rejects({});
         MeetingUtil.isPinOrGuest = sinon.stub().returns(true);
         const meeting = {
           id: 'meeting-id',
@@ -576,6 +616,8 @@ describe('plugin-meetings', () => {
               meetingId: meeting.id,
             },
           });
+        } finally {
+          joinMeetingSpy.restore();
         }
       });
     });